Project complete

+++ backend/routes/Blog.route.js
@@ -1,11 +1,19 @@
 const express = require('express');
 const BlogPost = require('../models/BlogModel');
 const router = express.Router();
+const { protect } = require('../middleware/authMiddleware');
 
-router.post('/', async (req, res) => {
+router.post('/', protect, async (req, res) => { 
   try {
-    const { title, content, author } = req.body;
-    const newPost = new BlogPost({ title, content, author });
+    const { title, content } = req.body;
+    
+    const newPost = new BlogPost({ 
+        title, 
+        content, 
+        authorId: req.user._id,
+        author: req.user.username 
+    });
+    
     const savedPost = await newPost.save();
     res.status(201).json(savedPost);
   } catch (error) {
@@ -37,32 +45,45 @@ router.get('/:id', async (req, res) => {
 });
 
 
-router.put('/:id', async (req, res) => {
+router.put('/:id', protect, async (req, res) => { 
   try {
-    const { title, content, author } = req.body;
-    const updatedPost = await BlogPost.findByIdAndUpdate(
-      req.params.id,
-      { title, content, author }, 
-      { new: true, runValidators: true } 
-    );
+    const post = await BlogPost.findById(req.params.id);
 
-    if (!updatedPost) {
-      return res.status(404).json({ message: 'Post not found' });
+    if (post) {
+      // Check if the logged-in user is the author of the post
+      if (post.authorId.toString() !== req.user._id.toString()) {
+        return res.status(403).json({ message: 'User not authorized to update this post' });
+      }
+
+      const updatedPost = await BlogPost.findByIdAndUpdate(
+        req.params.id,
+        req.body, 
+        { new: true, runValidators: true }
+      );
+      res.json(updatedPost);
+    } else {
+      res.status(404).json({ message: 'Post not found' });
     }
-    res.json(updatedPost);
   } catch (error) {
     res.status(400).json({ message: error.message });
   }
 });
 
 
-router.delete('/:id', async (req, res) => {
+router.delete('/:id', protect, async (req, res) => { 
   try {
-    const deletedPost = await BlogPost.findByIdAndDelete(req.params.id);
-    if (!deletedPost) {
-      return res.status(404).json({ message: 'Post not found' });
+    const post = await BlogPost.findById(req.params.id);
+
+    if (post) {
+      if (post.authorId.toString() !== req.user._id.toString()) {
+        return res.status(403).json({ message: 'User not authorized to delete this post' });
+      }
+
+      await BlogPost.findByIdAndDelete(req.params.id);
+      res.json({ message: 'Blog post deleted successfully' });
+    } else {
+      res.status(404).json({ message: 'Post not found' });
     }
-    res.json({ message: 'Blog post deleted successfully' });
   } catch (error) {
     res.status(500).json({ message: error.message });
   }
